refactor: game_jrpg 폴더 생성 및 파일 이동
lib 및 test 폴더의 파일들을 game_jrpg 하위 폴더로 이동하고, 관련 import 경로를 수정
