+++ lib/game_jrpg/slime.dart
