+++ test/game_jrpg/wand_test.dart
@@ -1,6 +1,6 @@
 import 'dart:math';
 
-import 'package:modu_3_dart_study/wand.dart';
+import 'package:modu_3_dart_study/game_jrpg/wand.dart';
 import 'package:test/expect.dart';
 import 'package:test/scaffolding.dart';
